airflow/requirements.txt
# Airflow providers and dependencies
apache-airflow-providers-docker==3.10.0
